+++ PlatformShared/Messages/AdminRenewAppMessage.cs
@@ -7,9 +7,9 @@ namespace PlatformShared.Messages
     {
         public required int Months { get; init; }
         public required string Comment { get; init; }
-        public required int ApproverLocalId { get; init; }
+        public required long ApproverLocalId { get; init; }
         public required int Requester { get; init; }
-        public required int RequesterLocalId { get; init; }
+        public required long RequesterLocalId { get; init; }
         public required string RequesterOrgName { get; init; }
     }
 
@@ -37,7 +37,7 @@ namespace PlatformShared.Messages
         /// Requester local id
         /// 请求人本地编号
         /// </summary>
-        public required int RequesterLocalId { get; init; }
+        public required long RequesterLocalId { get; init; }
 
         /// <summary>
         /// Requester organization id
@@ -61,7 +61,7 @@ namespace PlatformShared.Messages
         /// Approver local id
         /// 批准人本地编号
         /// </summary>
-        public required int ApproverLocalId { get; init; }
+        public required long ApproverLocalId { get; init; }
 
         /// <summary>
         /// Comment
